Create configuration package for tool configuration
Moving the configuration file handling code to a dedicated internal package will make it avaiable for use by any of the
module's packages.
