+++ internal/configuration/configuration.go
@@ -1,0 +1,64 @@
+// This file is part of libraries-repository-engine.
+//
+// Copyright 2021 ARDUINO SA (http://www.arduino.cc/)
+//
+// This program is free software: you can redistribute it and/or modify
+// it under the terms of the GNU Affero General Public License as published
+// by the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+//
+// This program is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU Affero General Public License for more details.
+//
+// You should have received a copy of the GNU Affero General Public License
+// along with this program.  If not, see <https://www.gnu.org/licenses/>.
+//
+// You can be released from the requirements of the above licenses by purchasing
+// a commercial license. Buying such a license is mandatory if you want to
+// modify or otherwise use the software for commercial activities involving the
+// Arduino software without disclosing the source code of your own applications.
+// To purchase a commercial license, send an email to license@arduino.cc.
+
+// Package configuration handles the tool configuration.
+package configuration
+
+import (
+	"encoding/json"
+	"os"
+
+	"github.com/arduino/libraries-repository-engine/internal/feedback"
+)
+
+// Config is the type of the engine configuration.
+type Config struct {
+	BaseDownloadURL string
+	LibrariesFolder string
+	LogsFolder      string
+	LibrariesDB     string
+	LibrariesIndex  string
+	GitClonesFolder string
+	DoNotRunClamav  bool
+	ArduinoLintPath string
+}
+
+// ReadConf reads the configuration file and returns the data.
+func ReadConf(configFile string) *Config {
+	if _, err := os.Stat(configFile); os.IsNotExist(err) {
+		feedback.LogError(err)
+		os.Exit(1)
+	}
+
+	file, err := os.Open(configFile)
+	if feedback.LogError(err) {
+		os.Exit(1)
+	}
+	decoder := json.NewDecoder(file)
+	config := Config{}
+	err = decoder.Decode(&config)
+	if feedback.LogError(err) {
+		os.Exit(1)
+	}
+	return &config
+}
